@@ -2,7 +2,7 @@ package org.vsanyc.sandbox.spring.core.post.processor;
 
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 
 @Component("notOrderedBean")
 public class NotOrderedBean implements BeanInterface {
