@@ -2,7 +2,8 @@ package org.vsanyc.sandbox.spring.core.post.processor;
 
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
+
 
 @Component("orderedBean")
 public class OrderedBean implements BeanInterface {
